refactor: TypeScript test setup
