+++ electron/babel-register.js
@@ -1,0 +1,16 @@
+require('@babel/register')({
+  cache: false,
+  extensions: ['.ts'],
+  plugins: ['@babel/proposal-class-properties'],
+  presets: [
+    [
+      '@babel/preset-env',
+      {
+        targets: {
+          node: 'current',
+        },
+      },
+    ],
+    '@babel/preset-typescript',
+  ],
+});
