--- electron/test/mocha.opts
@@ -1,4 +1,0 @@
-# https://mochajs.org/#mochaopts
---bail
---reporter spec
-test/main
